Introduce ECS api mediatior to list container instances
Created a trait to serve as wrapper to ecs api call and map response
to our own type. This allows us to easily mock this trait and write integrations tests.

Also, Introduced snafu for error formatting.

Testing:
Added a sample integration tests which mocks the EcsMediator trait
and calls start with mocked mediator