Fixes a bug where instances are stuck in Drain state
Previously, when waiting on instance to be ok fails after starting update, error
was returned and updater was exited. However, instance remained in
drained state even if it would have become Ok after some time. This
change, changes instance state back to active irrespective of success
or failure of wait.