Fixed ssm update command waiting for command output
Sometimes ssm update command will not report output because instance
goes into reboot before it can report output. In this change, we do not
try to solve the problem where SSM command output is not available, instead
we understand that it could happen and we handle it in code. We split
update command in two steps apply and reboot, and we do not wait for reboot command output
to be available. Instead we rely on veriyUpdate to report correct status of update.