Added timeout seconds for ssm command
If ECS execution role is deleted SSM command will not be delivered
and by default the delivery timeout time is 3600 seconds or 1hour.
This change adds a delivery timeout of 600 seconds or 10 min so that
waiters does not have to wait so long for ssm command to complete.
After this change ssm command timeouts after 10 min (timeout) +
30 min (execution timeout) = 40min.